% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_exam.R
\name{exam}
\alias{exam}
\title{Tutorial exam questions}
\usage{
exam(..., caption = rlang::missing_arg())
}
\arguments{
\item{...}{One or more questions or answers}

\item{caption}{Optional quiz caption (defaults to "Quiz")}
}
\value{
A learnr quiz, or collection of questions.
}
\description{
Add interactive exam questions to a tutorial. Each exam question is executed
within a shiny runtime to provide more flexibility in the types of questions
offered.
}
