% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_option.R
\name{grade_server}
\alias{grade_server}
\title{Tutorial grade server}
\usage{
grade_server(
  id,
  rubric_list,
  num_try = 3,
  deduction = 0.1,
  display = c("itemize", "scaled")
)
}
\arguments{
\item{id}{ID matching ui with server}

\item{rubric_list}{A data frame containing a vector of question names and a vector of points_possible for each question}

\item{num_try}{Number of tries allowed before grade deduction on that question. Default is 3.}

\item{deduction}{The percent (as a decimal) to be deducted for each additional incorrect attempt after num_try. Default is 0.1.}

\item{display}{vector of at least one c("percent", "itemize", "scaled"). Default is c("itemize", "scaled").}
}
\description{
Tutorial grade server
}
