% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_exam.R
\name{inner_lock_server}
\alias{inner_lock_server}
\title{Tutorial inner lock server}
\usage{
inner_lock_server(id)
}
\arguments{
\item{id}{ID matching ui with server}
}
\description{
Tutorial inner lock server
}
