% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_exam.R
\name{knit_print.exam}
\alias{knit_print.exam}
\title{Knitr quiz print methods}
\usage{
\method{knit_print}{exam}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
\code{knitr::\link[knitr]{knit_print}} methods for \code{\link{question}} and \code{\link{quiz}}
}
