% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_exam.R
\name{lock_exam_button}
\alias{lock_exam_button}
\title{Lock tutorial exam}
\usage{
lock_exam_button(id = "lock", label = "lock answers")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the button}
}
\description{
Lock all the exam answers for submission.
}
