% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lock_question.R
\name{lock_q_button}
\alias{lock_q_button}
\title{Lock tutorial exam}
\usage{
lock_q_button(id = "Q-1", label = "lock Q1")
}
\arguments{
\item{id}{ID matching ui with server}

\item{label}{Label to appear on the button}
}
\description{
Lock all the exam answers for submission.
}
