% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_quiz.R
\name{notes_prerendered_chunk}
\alias{notes_prerendered_chunk}
\title{function to print notes}
\usage{
notes_prerendered_chunk(question, ..., session = getDefaultReactiveDomain())
}
\description{
function to print notes
}
