% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_option.R
\name{print_ui}
\alias{print_ui}
\title{Tutorial print option}
\usage{
print_ui(id, label = "pdf")
}
\arguments{
\item{id}{ID to link to ui}

\item{label}{Label to appear on the print button}
}
\description{
Print the tutorial. Should load print-file.css into tutorial yaml
for nice looking document.
}
