% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_blanks.R
\name{question_blank}
\alias{question_blank}
\title{Fill in the blank question for learnr tutorials.}
\usage{
question_blank(
  text,
  ...,
  type = "blank",
  correct = paste0(icons::ionicons("checkmark-outline")),
  incorrect = paste0(icons::ionicons("close-outline")),
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE,
  placeholder = "Enter answer here...",
  trim = TRUE,
  rows = NULL,
  cols = NULL,
  options = list()
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{type}{Type of quiz question. Typically this can be automatically determined
based on the provided answers. Pass \code{"radio"} to indicate that even though
multiple correct answers are specified that inputs which include only one correct
answer are still correct. Pass \code{"checkbox"} to force the use of checkboxes
(as opposed to radio buttons) even though only once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{options}{Extra options to be stored in the question object.}
}
\value{
A custom \code{learnr} question, with \code{type = blank}.
}
\description{
Add interactive fill in the blank tasks to your \code{learnr} tutorials.
}
