% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_matching.R
\name{question_matching}
\alias{question_matching}
\title{Ranking question for learnr tutorials.}
\usage{
question_matching(
  text,
  ...,
  choices = choices,
  correct = "Correct!",
  incorrect = "Incorrect",
  loading = c("**Loading:** ", text, "<br/><br/><br/>"),
  submit_button = "Submit Answer",
  try_again_button = "Try Again",
  allow_retry = FALSE,
  random_answer_order = TRUE,
  options = sortable::sortable_options()
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{choices}{a vector of choices that will remain stationary that the answers are sorted with.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{loading}{Loading text to display as a placeholder while the question is loaded}

\item{submit_button}{Label for the submit button. Defaults to \code{"Submit Answer"}}

\item{try_again_button}{Label for the try again button. Defaults to \code{"Submit Answer"}}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{options}{Extra options to be stored in the question object.}
}
\value{
A custom \code{learnr} question, with \code{type = matching}.
}
\description{
Thank you to the sortable package for creating functions that allow ranking.
This is copied from the function question_rank to allow for matching options.
}
\details{
Add interactive ranking tasks to your \code{learnr} tutorials.  The student can
drag-and-drop the answer options into the desired order.

Each set of answer options must contain the same set of answer options. When
the question is completed, the first correct answer will be displayed.

Note that, by default, the answer order is randomized.
}
