% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_multidrop.R
\name{question_multidrop}
\alias{question_multidrop}
\title{Multiple dropdown question}
\usage{
question_multidrop(
  text,
  choices,
  ...,
  box = 6,
  type = "multidrop",
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE
)
}
\arguments{
\item{text}{Question or option text}

\item{choices}{a vector of choices that will remain stationary in the left column.}

\item{...}{parameters passed onto learnr answer.}

\item{box}{a number between 1 and 11, inclusive, indicating the width of the dropdown box.
The default is 6 which corresponds to 50\% of the page width.}

\item{type}{Type of quiz question. Typically this can be automatically determined
based on the provided answers. Pass \code{"radio"} to indicate that even though
multiple correct answers are specified that inputs which include only one correct
answer are still correct. Pass \code{"checkbox"} to force the use of checkboxes
(as opposed to radio buttons) even though only once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}
}
\value{
A custom \code{learnr} question, with \code{type = multidrop}.
}
\description{
Creates a multiple dropdown tutorial quiz question.
Each dropdown will have the same list of options that correspond to a vector of questions.
Each dropdown only allow only one selection.
}
\seealso{
Other Interactive Questions: 
\code{\link{quiz_notes}()}
}
\concept{Interactive Questions}
