% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_notes.R
\name{quiz}
\alias{quiz}
\alias{quiz_notes}
\title{Tutorial notes format}
\usage{
quiz_notes(..., caption = rlang::missing_arg())
}
\arguments{
\item{...}{One or more questions or answers}

\item{caption}{Optional quiz caption (defaults to "Quiz")}
}
\description{
Add interactive notes to a tutorial.
}
\concept{Interactive Questions}
