% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_notes.R
\name{quiz_notes}
\alias{quiz_notes}
\title{Tutorial notes format}
\usage{
quiz_notes(..., caption = rlang::missing_arg())
}
\arguments{
\item{...}{One or more questions or answers}

\item{caption}{Optional quiz caption (defaults to "Quiz")}
}
\description{
Add interactive notes to a tutorial. This is an alternative format to
display fill in the blank lines with smaller submit buttons for a
more fluid document then the boxed questions. This was adapted from
the learnr package quiz function. Many thanks to the learnr author
who developed a great and adaptable teaching tool.
}
\seealso{
quiz

Other Interactive Questions: 
\code{\link{exam}()}
}
\concept{Interactive Questions}
