% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam}
\alias{exam}
\title{Exam container}
\usage{
exam(..., caption = rlang::missing_arg())
}
\arguments{
\item{...}{One or more questions or answers}

\item{caption}{Optional quiz caption (defaults to "Quiz")}
}
\description{
Add an exam format to prevent displaying whether the question was correct
or incorrect. This was adapted from the learnr package quiz
function. Many thanks to the learnr author who developed a great and adaptable teaching tool.
}
\seealso{
quiz

Other Interactive Questions: 
\code{\link{quiz_notes}()}
}
\concept{Interactive Questions}
