% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_blanks.R
\name{question_blank}
\alias{question_blank}
\title{Fill in the blank question for learnr tutorials.}
\usage{
question_blank(
  text,
  ...,
  type = "blank",
  correct = paste0(icons::ionicons("checkmark-outline")),
  incorrect = paste0(icons::ionicons("close-outline")),
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE,
  placeholder = "Enter answer here...",
  trim = TRUE,
  rows = NULL,
  cols = NULL,
  options = list()
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{\ifelse{{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}} Random answer
order for text questions is automatically disabled to ensure that the
submission is checked against each answer in the order they were provided
by the author.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{trim}{Logical to determine if whitespace before and after the answer
should be removed.  Defaults to \code{TRUE}.}

\item{rows}{Defines the size of the text input area in terms of the
number of rows or character columns visible to the user. If either \code{rows}
or \code{cols} are provided, the quiz input will use \code{\link[shiny:textAreaInput]{shiny::textAreaInput()}}
for the text input, otherwise the default input element is a single-line
\code{\link[shiny:textInput]{shiny::textInput()}}.}

\item{cols}{Defines the size of the text input area in terms of the
number of rows or character columns visible to the user. If either \code{rows}
or \code{cols} are provided, the quiz input will use \code{\link[shiny:textAreaInput]{shiny::textAreaInput()}}
for the text input, otherwise the default input element is a single-line
\code{\link[shiny:textInput]{shiny::textInput()}}.}

\item{options}{Extra options to be stored in the question object.}
}
\value{
A custom \code{learnr} question, with \code{type = blank}.
}
\description{
Add interactive fill in the blank tasks to your \code{learnr} tutorials.
}
