% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_blanks.R
\name{question_blank}
\alias{question_blank}
\title{Text blank question}
\usage{
question_blank(
  text,
  ...,
  type = "learnr_blank",
  correct = paste0(icons::ionicons("checkmark-outline")),
  incorrect = paste0(icons::ionicons("close-outline")),
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE,
  placeholder = "Enter answer here...",
  trim = TRUE,
  rows = NULL,
  cols = NULL,
  options = list()
)
}
\arguments{
\item{...}{Answers created with \code{\link[=answer]{answer()}} or \code{\link[=answer_fn]{answer_fn()}}, or extra
parameters passed onto \code{\link[=question]{question()}}. Answers with custom function checking}

\item{random_answer_order}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Random answer
order for text questions is automatically disabled to ensure that the
submission is checked against each answer in the order they were provided
by the author.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{trim}{Logical to determine if whitespace before and after the answer
should be removed.  Defaults to \code{TRUE}.}

\item{rows, cols}{Defines the size of the text input area in terms of the
number of rows or character columns visible to the user. If either \code{rows}
or \code{cols} are provided, the quiz input will use \code{\link[shiny:textAreaInput]{shiny::textAreaInput()}}
for the text input, otherwise the default input element is a single-line
\code{\link[shiny:textInput]{shiny::textInput()}}.}
}
\value{
Returns a learnr question of type \code{"learnr_blank"}.
}
\description{
Creates a tutorial question asking the student to enter text. The default
text input is appropriate for short or single-line text entry. For longer
text input, set the \code{rows} and/or \code{cols} argument to create a larger text
area.

When used with \code{\link[=answer]{answer()}}, the student's submission must match the answer
exactly, minus whitespace trimming if enabled with \code{trim = TRUE}. For more
complicated submission evaluation, use \code{\link[=answer_fn]{answer_fn()}} to provide a function
that checks the student's submission. For example, you could provide a
function that evaluates the user's submission using
\link[base:regex]{regular expressions}.
}
\seealso{
Other Interactive Questions: 
\code{\link{question_dropdown}()}
}
\concept{Interactive Questions}
