% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_dropdown.R
\name{question_dropdown}
\alias{question_dropdown}
\title{Dropdown question for learnr tutorials.}
\usage{
question_dropdown(
  text,
  ...,
  type = "dropdown",
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto \code{\link[learnr:quiz]{learnr::question()}}.}

\item{type}{Type of quiz question. Typically this can be automatically determined
based on the provided answers. Pass \code{"radio"} to indicate that even though
multiple correct answers are specified that inputs which include only one correct
answer are still correct. Pass \code{"checkbox"} to force the use of checkboxes
(as opposed to radio buttons) even though only once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}
}
\value{
A custom \code{learnr} question, with \code{type = dropdown}.
See \code{\link[learnr:quiz]{learnr::question()}}.

Returns a learnr question of type \code{"dropdown"}.
}
\description{
Add interactive drop down tasks to your \code{learnr} tutorials.
}
