% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dropdown.R
\name{question_dropdown}
\alias{question_dropdown}
\title{Dropdown question}
\usage{
question_dropdown(
  text,
  ...,
  type = "learnr_dropdown",
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  random_answer_order = FALSE
)
}
\arguments{
\item{...}{Answers created with \code{\link[=answer]{answer()}} or extra parameters passed onto
\code{\link[=question]{question()}}. Function answers are ignored for radio questions because the
user is required to select a single answer.}
}
\value{
Returns a learnr question of type \code{"learnr_dropdown"}.
}
\description{
Creates a dropdown button tutorial quiz question. The student can select only
one dropdown before submitting their answer. Note: Multiple correct
answers are allowed.
}
\examples{
question_dropdown(
  "Pick the letter B",
  answer("A"),
  answer("B", correct = TRUE),
  answer("C"),
  answer("D"),
  allow_retry = TRUE,
  random_answer_order = TRUE
)

}
\seealso{
Other Interactive Questions: 
\code{\link{question_blank}()}
}
\concept{Interactive Questions}
