% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_matching.R
\name{question_matching}
\alias{question_matching}
\title{Matching order question for learnr tutorials}
\usage{
question_matching(
  text,
  ...,
  choices = choices,
  correct = "Correct!",
  incorrect = "Incorrect",
  loading = c("**Loading:** ", text, "<br/><br/><br/>"),
  submit_button = "Submit Answer",
  try_again_button = "Try Again",
  allow_retry = FALSE,
  random_answer_order = TRUE,
  style = "tutorial_question",
  options = sortable::sortable_options()
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{choices}{a vector of choices that will remain stationary in the left column.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{loading}{Loading text to display as a placeholder while the question is loaded}

\item{submit_button}{Label for the submit button. Defaults to \code{"Submit Answer"}}

\item{try_again_button}{Label for the try again button. Defaults to \code{"Submit Answer"}}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{style}{can change display of question to "notes" or "exam"}

\item{options}{Extra options to be stored in the question object.}
}
\value{
A custom \code{learnr} question, with \code{type = matching}.
}
\description{
The following function was modified from the sortable package, available at
https://github.com/rstudio/sortable. Many thanks to the sortable author
who developed tools to drag and drop objects in rank order. This extends
the "question_rank" option so that you can match the options in the right
column with the choices on the left.
}
\details{
Add interactive matching tasks to your \code{learnr} tutorials.  The student can
drag-and-drop the answer options in the right column to match the corresponding
options in the left column.

The choices options should be a set of static options that will appear in the
left column. It should be of equal length to the answer options.

The answer options should contain the set of options that need to be rearranged
in the right column. The answer set should correspond with the correct order for
the choices option. The display order will be automatically shuffled for the user.
}
\examples{
question_matching(
  "Rearrange the numbers to match the corresponding letter's position in the alphabet.",
  choices = c("A", "B", "C", "D"),
  learnr::answer(c("1", "2", "3", "4"), correct = TRUE),
  allow_retry = TRUE
)

}
