% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_multidrop.R
\name{question_multidrop}
\alias{question_multidrop}
\title{Multiple drop-down question for learnr tutorials}
\usage{
question_multidrop(
  text,
  ...,
  choices,
  wordbank = NULL,
  box = 6,
  arrange = "random",
  type = "multidrop",
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  allow_retry = FALSE,
  style = "tutorial_question",
  random_answer_order = FALSE
)
}
\arguments{
\item{text}{Question or option text}

\item{...}{parameters passed onto learnr answer.}

\item{choices}{a vector of questions/options that will remain stationary in the right column.}

\item{wordbank}{a vector of drop-down options when providing more options than answers.
If NULL then wordbank will be set equal to the answer choices.}

\item{box}{a number between 1 and 11, inclusive, indicating the width of the drop-down box.
The default is 6 which corresponds to 50\% of the page width.}

\item{arrange}{either 'random' or 'ordered'; default is random. Set equal to ordered if
you want the drop-down list to appear alphabetically.}

\item{type}{Type of quiz question. Typically this can be automatically determined
based on the provided answers. Pass \code{"radio"} to indicate that even though
multiple correct answers are specified that inputs which include only one correct
answer are still correct. Pass \code{"checkbox"} to force the use of checkboxes
(as opposed to radio buttons) even though only once correct answer was provided.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{style}{can change display of question to "notes" or "exam"}

\item{random_answer_order}{Display answers in a random order.}
}
\value{
A custom \code{learnr} question, with \code{type = multidrop}.
}
\description{
Creates a multiple drop-down tutorial quiz question.
Each drop-down will have the same list of options that correspond to a vector of questions.
Each drop-down only allow only one selection.
}
\examples{
question_multidrop(
  "True or False",
  choices = c("1+1 = 2", "4*2 =16", "7-5 =2"),
  learnr::answer(c("True", "False", "True"), correct = TRUE),
  allow_retry = TRUE
)

}
