% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_question.R
\name{drop_question}
\alias{drop_question}
\alias{blank_question}
\alias{quiz}
\alias{quiz_notes}
\title{Tutorial quiz questions}
\usage{
drop_question(
  text,
  ...,
  type = c("learnr_dropdown"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", format(text), "<br/><br/><br/>"),
  submit_button = rlang::missing_arg(),
  try_again_button = rlang::missing_arg(),
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)

blank_question(
  text,
  ...,
  type = c("learnr_blank"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", format(text), "<br/><br/><br/>"),
  submit_button = rlang::missing_arg(),
  try_again_button = rlang::missing_arg(),
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)

quiz_notes(..., caption = rlang::missing_arg())
}
\arguments{
\item{...}{One or more questions or answers}

\item{type}{Type of quiz question.}

\item{correct}{For \code{question}, text to print for a correct answer (defaults
to "Correct!"). For \code{answer}, a boolean indicating whether this answer is
correct.}

\item{incorrect}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{FALSE}.}

\item{try_again}{Text to print for an incorrect answer (defaults to "Incorrect")
when \code{allow_retry} is \code{TRUE}.}

\item{message}{Additional message to display along with correct/incorrect feedback.
This message is always displayed after a question submission.}

\item{post_message}{Additional message to display along with correct/incorrect feedback.
If \code{allow_retry} is \code{TRUE}, this message will only be displayed after the
correct submission.  If \code{allow_retry} is \code{FALSE}, it will produce a second
message alongside the \code{message} message value.}

\item{loading}{Loading text to display as a placeholder while the question is loaded}

\item{submit_button}{Label for the submit button. Defaults to \code{"Submit Answer"}}

\item{try_again_button}{Label for the try again button. Defaults to \code{"Submit Answer"}}

\item{allow_retry}{Allow retry for incorrect answers. Defaults to \code{FALSE}.}

\item{random_answer_order}{Display answers in a random order.}

\item{options}{Extra options to be stored in the question object.}

\item{caption}{Optional quiz caption (defaults to "Quiz")}
}
\description{
Add interactive quiz questions to a tutorial.
Each quiz question is executed within a shiny runtime to provide more flexibility in the types of questions offered.
There are two new types of quiz questions:
\describe{
\item{\code{learnr_blank}}
\item{\code{learnr_dropdown}}
}

Add interactive notes to a tutorial.
}
\seealso{
\code{\link[=random_praise]{random_praise()}}, \code{\link[=random_encouragement]{random_encouragement()}}

Other Interactive Questions: 
\code{\link{question_blank}()},
\code{\link{question_dropdown}()}
}
\concept{Interactive Questions}
