% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_blanks.R, R/question_dropdown.R,
%   R/question_matching.R, R/quiz_notes.R
\name{blank_question}
\alias{blank_question}
\alias{drop_question}
\alias{matching_question}
\alias{quiz}
\alias{quiz_notes}
\title{Tutorial notes format}
\usage{
blank_question(
  text,
  ...,
  type = c("blank"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", format(text), "<br/><br/><br/>"),
  submit_button = rlang::missing_arg(),
  try_again_button = rlang::missing_arg(),
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)

drop_question(
  text,
  ...,
  type = c("dropdown"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", format(text), "<br/><br/><br/>"),
  submit_button = rlang::missing_arg(),
  try_again_button = rlang::missing_arg(),
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)

matching_question(
  text,
  ...,
  choices = choices,
  type = c("matching"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", format(text), "<br/><br/><br/>"),
  submit_button = rlang::missing_arg(),
  try_again_button = rlang::missing_arg(),
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)

quiz_notes(..., caption = rlang::missing_arg())
}
\arguments{
\item{...}{One or more questions or answers}

\item{caption}{Optional quiz caption (defaults to "Quiz")}
}
\description{
Add interactive notes to a tutorial.
}
\concept{Interactive Questions}
